+++ api/src/parser/terms/term_parser.rs
@@ -7,7 +7,8 @@ use pest_derive::Parser;
 pub(crate) struct TermParser;
 
 #[cfg(test)]
-#[allow(clippy::cognitive_complexity)]
+// Allowing manual_assert because of pest's `parses_to` macro.
+#[allow(clippy::cognitive_complexity, clippy::manual_assert)]
 mod tests {
     use super::{TermParser, *};
     use pest::{consumes_to, parses_to, Parser};