+++ atom_generator/src/toml_structs/toml_atom.rs
@@ -1,4 +1,4 @@
-use heck::CamelCase;
+use heck::ToUpperCamelCase;
 use regex::Regex;
 
 lazy_static! {
@@ -57,23 +57,23 @@ fn sanitize_type_name_segment(mut string: String) -> String {
 
     let s = remove_non_letter_chars(&string);
 
-    s.to_camel_case()
+    s.to_upper_camel_case()
 }
 
 #[allow(clippy::non_ascii_literal)]
 fn remove_non_latin_chars(string: &str) -> String {
     string
-        .replace("è", "e")
-        .replace("é", "e")
-        .replace("Å", "A")
-        .replace("ö", "o")
+        .replace('è', "e")
+        .replace('é', "e")
+        .replace('Å', "A")
+        .replace('ö', "o")
 }
 
 fn remove_non_letter_chars(string: &str) -> String {
     string
-        .replace("'", "")
-        .replace("*", " Star")
-        .replace("^", " Caret")
+        .replace('\'', "")
+        .replace('*', " Star")
+        .replace('^', " Caret")
 }
 
 fn finally(name: &str, primary_code: &str) -> String {