+++ api/CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.22.0] - 2022-03-23
+
+### Added
+
+- Implemented `AsRef<Self>` for `Measurement` and `Unit`. Allows for downstream wrapper types to
+  implement other functionality for all types via the `AsRef` implementation.
+- Implemented `TryFrom<&str> for Unit` (which just calls `from_str()`) to allow for downstream
+  wrapper implementations around `Unit`.
+- New `const` `Composition` methods: `new_dimless()`, `new_any()`.
+- New `composition` `const`s for common dimensional compositions.
+
+### Changed
+
+- Exported `parser::composition` from the crate root.
+- `Measurement::try_new()` now takes `unit: U`, `where Unit: TryFrom<U, crate::Error>`. This still
+  allows for passing in a unit expression as a `&str`, but also allows for any other
+  implementations of type conversions to `Unit`.
+- Changed `AsFraction` trait definition to allow for non-optional `Numerator` and `Denominator`.
+- Updated to Rust edition `2021`.
+
 ## [0.21.1] - 2021-11-19
 
 ### Changed